refactor: move Config and Action into config module
