+++ src/main.rs
@@ -1,15 +1,8 @@
-use std::{env, error::Error, process};
-
-use time_tracker::Config;
+use std::{env, error::Error};
 
 fn main() -> Result<(), Box<dyn Error>> {
     let args: Vec<String> = env::args().collect();
-    let config = Config::build(&args).unwrap_or_else(|err| {
-        eprintln!("Problem parsing arguments: {err}");
-        process::exit(1);
-    });
-
-    time_tracker::run(config)?;
+    time_tracker::run(&args)?;
     // if let Err(e) = time_tracker::run(config) {
     //     eprintln!("Application error: {e}");
     //     process::exit(1);
